fix(logger): Prevent infinite recursion with log.condition => matches
When we need to check the log condition for a user matches,

there is a risk that something on the way checks the log level

and would result in an infinite loop.

So we simply check if it's a nested call and use the default

warning level in that case.

Signed-off-by: Joas Schilling <coding@schilljs.com>

[skip ci]